+++ onmt/utils/init.lua
@@ -15,5 +15,6 @@ utils.Logger = require('onmt.utils.Logger')
 utils.Profiler = require('onmt.utils.Profiler')
 utils.ExtendedCmdLine = require('onmt.utils.ExtendedCmdLine')
 utils.CrayonLogger = require('onmt.utils.CrayonLogger')
+utils.Error = require('onmt.utils.Error')
 
 return utils