+++ tools/score.lua
@@ -8,8 +8,7 @@ local cmd = onmt.utils.ExtendedCmdLine.new('scorer.lua')
 
 local options = {
   {
-    '-scorer',
-    onmt.scorers.list[1],
+    '-scorer', 'bleu',
     [[Scorer to use.]],
     {
       enum = onmt.scorers.list
@@ -25,8 +24,8 @@ local options = {
   },
   {
     '-sample',
-    10,
-    [[Number of sample for estimation of error margin.]],
+    1,
+    [[If > 1, number of samples for estimation of k-fold error margin (95% certitude) - 10 is a good value.]],
     {
       valid = onmt.utils.ExtendedCmdLine.isUInt()
     }
@@ -55,7 +54,7 @@ local function main()
 
   local function add_to_reference(filename)
     local file = io.open(filename)
-    assert(file, "cannot open `" .. filename .. "`")
+    onmt.utils.Error.assert(file, "cannot open `" .. filename .. "`")
     local ref = {}
     while true do
       local line = file:read()
@@ -66,7 +65,7 @@ local function main()
       end
       table.insert(ref, sent)
     end
-    assert(#references==0 or #references[#references] == #ref, "ERROR: all references do not have same line count")
+    onmt.utils.Error.assert(#references==0 or #references[#references] == #ref, "all references do not have same line count")
     table.insert(references, ref)
   end
 
@@ -76,9 +75,11 @@ local function main()
     add_to_reference(opt.rfilestem)
   end
 
-  while path.exists(opt.rfilestem .. refid) do
-    add_to_reference(opt.rfilestem .. refid)
-    refid = refid + 1
+  if onmt.scorers.multi[opt.scorer] then
+    while path.exists(opt.rfilestem .. refid) do
+      add_to_reference(opt.rfilestem .. refid)
+      refid = refid + 1
+    end
   end
 
   local hyp = {}
@@ -93,13 +94,38 @@ local function main()
     table.insert(hyp, sent)
   end
 
-  assert(#hyp==#references[1], "ERROR: line count hyp/ref does not match")
+  onmt.utils.Error.assert(#hyp==#references[1], "line count hyp/ref does not match")
 
-  _G.logger:info("%d references, %d sentences", #references, #hyp)
+  if not onmt.scorers.multi[opt.scorer] then
+    references = references[1]
+  end
 
-  local details = select(2, onmt.scorers[opt.scorer](hyp, references, opt.sample, opt.order))
-  print(details)
+  local score, format = onmt.scorers[opt.scorer](hyp, references, opt.order)
+  score = string.format("%.2f", score*100)
+  local margin = ''
 
+  if opt.sample > 1 then
+    local scores = torch.Tensor(opt.sample)
+    for k = 1, opt.sample do
+      -- extract 1/2 random sample
+      local perm = torch.randperm(#hyp)
+      local nhyp = {}
+      local nref = {}
+      for _ = 1, #references do
+        table.insert(nref, {})
+      end
+      for i = 1, #hyp/2 do
+        table.insert(nhyp, hyp[perm[i]])
+        for j = 1, #references do
+          table.insert(nref[j], references[j][perm[i]])
+        end
+      end
+      scores[k] = onmt.scorers[opt.scorer](nhyp, nref, opt.order)
+    end
+    score = string.format("%.2f", torch.mean(scores)*100)
+    margin = string.format("+/-%.2f", torch.std(scores)*1.96*100)
+  end
+  print(score, margin, format)
 end
 
 main()