+++ CHANGELOG.md
@@ -2,6 +2,32 @@
 
 ### Breaking changes
 
+* Remove tokenization options from `learn_bpe.lua`
+* Learning rate is also decayed when using Adam
+
+### New features
+
+* Add TER calculation
+* Add CNN encoder
+* Add option to save the validation translation to a file
+* Add 'Shallow Fusion' of language model in decoder
+* Add option to reset the optimizer states when the learning rate is decayed
+* Add option to dump attention in `translate.lua`
+
+### Fixes and improvements
+
+* Allow disabling gradients clipping with `-max_grad_norm 0`
+* Allow disabling global parameters initialization with `-param_init 0`
+* Introduce error estimation in scorer for all metrics
+* Fix validation score delta that was applied in the incorrect direction
+* Fix LuaJIT out of memory issues in `learn_bpe.lua`
+* More graceful error handling
+* Reduce memory footprint of Adam, Adadelta and Adagrad optimizers
+
+## [v0.8.0](https://github.com/OpenNMT/OpenNMT/releases/tag/v0.8.0) (2017-06-28)
+
+### Breaking changes
+
 * Models previously trained with `-pdbrnn` or `-dbrnn` are no more compatible
 * `-start_decay_ppl_delta` option is renamed to `-start_decay_score_delta`
 * `-decay perplexity_only` option is renamed to `-decay score_only`
@@ -9,18 +35,20 @@
 ### Deprecations
 
 * `-brnn`, `-dbrnn` and `-pdbrnn` options are replaced by `-encoder_type <type>` for future extensions
+* `-sample_tgt_vocab` option is renamed `-sample_vocab` and is extended to language models
 
 ### New features
 
-* Implement inference for language models
+* Implement inference for language models for scoring or sampling
 * Support variational dropout and dropout on source sequence
-* Support several validation metrics: loss, perplexity and BLEU
+* Support several validation metrics: loss, perplexity, BLEU and Damerau-Levenshtein edit ratio
 * Add option in preprocessing to check that lengths of source and target are equal (e.g. for sequence tagging)
 * Add `-pdbrnn_merge` option to define how to reduce the time dimension
 * Add option to segment mixed cased words
 * Add option to segment words of given alphabets or when switching alphabets
 * Add Google's NMT encoder
-* Add scorer tool (BLEU supported)
+* Add external scorer script for BLEU and Damerau-Levenshtein edit ratio
+* Add script to average multiple models
 * Add option to save the beam search as JSON
 
 ### Fixes and improvements